Refactor feature flag methods into a new module
Now that the implementations of add_/clear_feature_flag(s) are the same,
we can use a module to implement these methods and save us from having
copy pasted methods/docs in 3 places
