+++ lib/bugsnag/configuration.rb
@@ -18,6 +18,8 @@ require "bugsnag/endpoint_validator"
 
 module Bugsnag
   class Configuration
+    include Utility::FeatureDataStore
+
     # Your Integration API Key
     # @return [String, nil]
     attr_accessor :api_key
@@ -670,42 +672,6 @@ module Bugsnag
     # @api private
     def feature_flag_delegate
       request_data[:feature_flag_delegate] ||= Bugsnag::Utility::FeatureFlagDelegate.new
-    end
-
-    # Add a feature flag with the given name & variant
-    #
-    # @param name [String]
-    # @param variant [String, nil]
-    # @return [void]
-    def add_feature_flag(name, variant = nil)
-      feature_flag_delegate.add(name, variant)
-    end
-
-    # Merge the given array of FeatureFlag instances into the stored feature
-    # flags
-    #
-    # New flags will be appended to the array. Flags with the same name will be
-    # overwritten, but their position in the array will not change
-    #
-    # @param feature_flags [Array<Bugsnag::FeatureFlag>]
-    # @return [void]
-    def add_feature_flags(feature_flags)
-      feature_flag_delegate.merge(feature_flags)
-    end
-
-    # Remove the stored flag with the given name
-    #
-    # @param name [String]
-    # @return [void]
-    def clear_feature_flag(name)
-      feature_flag_delegate.remove(name)
-    end
-
-    # Remove all the stored flags
-    #
-    # @return [void]
-    def clear_feature_flags
-      feature_flag_delegate.clear
     end
 
     ##
